Move toJSON polyfill for BigInt from main.tsx into utils
