+++ src/types/global.d.ts
@@ -1,0 +1,8 @@
+declare global {
+  interface BigInt {
+    /** Convert BigInt to string form in JSON.stringify */
+    toJSON: () => string;
+  }
+}
+
+export {};
