+++ src/utils/polyfills.ts
@@ -1,0 +1,3 @@
+BigInt.prototype.toJSON = function () {
+  return this.toString();
+};
